Get set of meetings for a time range based on schedule

The `meetings` object consist of a schedule of AA meetings and locations in Iowa. This function creates a data frame of all available meetings and locations for a specified time range.

get_meetings(from = now(), to = now() + days(1),
  type = c("Alcoholics Anonymous", "Narcotics Anonymous"))

Arguments

from
earliest date (and time, if specified)

to
latest date

type
type of meeting

Value

data frame

Examples

library(leaflet)
library(sf)

# AA meetings in Iowa for the next 24h from
lubridate::now()

#> [1] "2020-07-09 14:24:25 CDT"

# color locations by day
pal <- colorFactor(
         palette = RColorBrewer::brewer.pal(n=7, name="Paired"),
         domain = levels(meetings$Day)
       )
get_meetings() %>%
  leaflet() %>%
  addTiles() %>%
  addPolygons(data = st_transform(ia_counties, crs='+proj=longlat +datum=WGS84'),
              weight = 1, color="#888888") %>%
  addCircleMarkers(lng = ~lon, lat = ~lat, color=~pal(Day),
                   label=~Meeting,
                   radius = 1, stroke = 0.1) %>%
  addLegend(pal = pal, values = levels(meetings$Day))