Iowa Hospital Buildings

Iowa hospital buildings listed in the federal Geographic Names Information System.

hospital_buildings

Format

A data frame with 1709 rows and 11 variables:

Feature.ID

Unique identifier for each record

Feature.Name

Name of building

Feature.Class

Type of building. All records in this dataset will be "hospital".

Primary.State

State of the location

Primary.County.Name

County of the location

Elevation..Meters.

Elevation in meters above sea level of the location

Elevation..Feet.

Elevation in feet above sea level of the location

USGS.Map.Name

Name of the location of the building

Date.Created

Date that the feature was added to the database

Date.Edited

Date at which the record was last updated

geometry

sf point object of geographic location

Source

https://data.iowa.gov/Physical-Geography/Iowa-Hospital-Buildings/ciqq-2z9p

Examples

library(dplyr)

# Leaflet map of hospital buildings in Iowa
library(leaflet)
library(sf)

hospital_buildings %>%
 leaflet() %>%
 addTiles() %>%
 addPolygons(data = ia_counties,
             weight = 1, color="#333333") %>%
 addCircleMarkers(radius = 1, stroke = 0.1)