South Central Behavioral Health Region resource directory

This dataset includes all resources in the resource directory of the South Central Behavioral Health Region of the Iowa MHDS regions. The dataset was downloaded and geocoded by Matthew Voss in July 2020.

scbhr_mhds

Format

A data frame with 170 rows and 10 variables:

category

the type of service provided

agency

the agency which offers the resource

address

the street address of the location

phone

the phone number for the resource

description

a description of the resource

website

facility website link

email

an email contact for the resource

last updated

date at which the information for the facility was last updated

longitude

geographic longitude

latitude

geographic latitude

Source

http://scbhr.org/resource-connection-2/