Childcare Providers

This dataset was scraped by Masoud Nosrati in May 2020. The data includes locations in Iowa of childcare providers The dataset was geocoded by Matthew Voss in August 2020.

childcare

Format

A data frame with 4335 rows and 22 variables:

dataset

name of the dataset

classification

type of services offered by daatset

county

name of the county

community

name of the community

provider_type

type of childcare provider

dhs_provider_id

id number of the childcare provider

provider_business_name

name of the childcare provider

provider_last_name

last name of the business owner

provider_first_name

first name of the business owner

is_active_ccaprovider

is the childcare provider a cca provider

address_1

address of the provider's location

address_2

additional address information of the provider's location

zip

zip code for the provider's location

phone

phone number of the provider

lic/reg_issue_date

the date of the provider's license issuance

lic/reg_expire_date

the date of the provider's license expiration

capacity

the capacity of the childcare provider

qrs_rating

unknown

ccrr_website

Iowa Child Care Resource and Referral

search_address

search address used in geocoding

longitude

geographic longitude

latitude

geographic latitude