Iowa Mental Health and Disability Services

This dataset includes mental health and disability service regional access points in Iowa. The dataset was collected and geocoded by Matthew Voss in July 2020. It includes access points for 11 of the 14 MHDS regions.

regional_MHDS

Format

A data frame with 79 rows and 15 variables:

classification

the type of service provided

dataset

the dataset in which the observation is

regional_org

the name of the MHDS region of the observation

source

the website that has information about the location

address

the street address of the location

unit_num

the unit number of the location, if it exists

city

the city of the location

state

the state of the location

zip

the five-digit zip code of the location

phone

the location's phone number

longitude

geographic longitude

latitude

geographic latitude

street_address

the full street address used for geocoding

name

the name of the location

Source

sources can be found in the dataset column "source"